feat: visualization modules
