+++ visualization/visualization_results/latest/summary.txt
@@ -1,5 +1,5 @@
 === SUMMARY STATISTICS ===
-Input file: results/latest\top_organisms_results.json
+Input file: ../results/latest\top_organisms_results.json
 Timestamp: 20250413_033956
 Most frequent top-ranking organism: NC_001348.1 Human herpesvirus 3, complete genome
 Best NRC value: 0.023945
